backend/requirements.txt
fiftyone
fastapi
uvicorn
python-multipart
sqlmodel
python-dotenv
pillow
numpy
requests
